feat: suggest corrections for unbound identifiers/fields based on distance metric  (#4720)
Add flag --ai-errors to tailor error reporting for human (default) or artificial users.

For humans, instead of reporting that a value or type identifier is not available in references or projections, we now suggest  likely candidates based on:

* whether the identifier is a prefix of the candidate identifiers.
* whether the identifier is some [edit distance](https://en.wikipedia.org/wiki/Levenshtein_distance) away from the candidate identifiers. 

and report identifiers in order of edit distance.

When the object/actor/module  projected from has few bindings (<16), we display them with their types, to avoid producing
huge error messages.

With --ai-errors, we don't give hints but display all available bindings with their types

(Thanks to Claude.ai for coughing up an implementation.)