chore: test recovery procedure when `finally` traps (#4841)
This PR adds a substitution variable `$PRINCIPAL` to `.drun` scripts, encoding a message argument `{ canister_id = $ID }`.
Using this information we can invoke the management canister's `start_canister`/`stop_canister` methods and with this we can implement a test for the recovery sequence detailed in #4840, to upgrade even if there are continuations in the futures table, e.g. due to trapping `finally` block.


Fixes #4840.