feat: add `.values()` alias for `.vals()` (#4876)
Makes it possible to write `x.values()` in place of `x.vals()` for an array or `Blob`. This is motivated by the upcoming base library changes in [this repository](https://github.com/dfinity/new-motoko-base).

`values()` is used in place of `vals()` for a large number of popular PLs (JS, Python, Java, C#, Rust, Swift, Kotlin, etc.), so this change would enable developers and LLMs to better transfer knowledge from other languages.

**Progress:**

* [x] Add alias in type checker
* [x] Update documentation and examples
* [x] Update unit tests to cover both `.vals()` and `.values()`