refactor: parentheticals KISS (#4900)
## KISS: keep it simple, stupid

This refactors the lowering of parentheticals for sends to simply assign to backend variables. So we piggy-back on the imperative mechanism instead of running the parenthetical through all the IR and backend hoops.

Maybe later we have to remove this shortcut, but for now it should work.

Thanks to @crusso for suggesting!

---------

Co-authored-by: Claudio Russo <claudio@dfinity.org>