chore: use the stock `js_of_ocaml` (#4965)
We used to build v5.0.1 ourselves, as there were correctness problems with v4.1.x. This reverts back to the supplied one that is bundled with OCaml 4.13.

------------
there is some fallout, though, as the `limit.mo` test won't let us to iterate up to 10000. I had to reduce it to the 7600 mark due to a stack overflow:
``` diff
>   {
    result: {
+     error: {}
-     error: null
    },
+   stderr: 'limit.mo:1.23-1.27: internal error, Stack overflow\n' +
+     '\n' +
+     'Last environment:\n' +
+     '@ManagementCanister = {}\n' +
+     '@add_cycles = <func>\n' +
```

I am trying to figure out where this started happening in #4969.