chore: bump `ic` to label `release-2025-02-27_03-09-base` (#4951)
This unlocks a few `drun` features (such as memory limit triggers). (@luc-blaeser Correct it if this is wrong!)

Needed some time to iron out the error message changes from the replica...

Also removed the CI step "pre-build `drun`", as we now have sufficient free room on the server.

----------------
_Resolved_: needs `EXTRA_MOC_ARGS="--enhanced-orthogonal-persistence"`

We have a problem with `.drun` tests:
```
(cd test/run-drun && ../run.sh -d any-upgrades.drun)
```
won't detect that the `.ok` file needs to be updated. In fact, passing `-da` it deletes the `.ok` file altogether!