feat: incremental partition growth for the incremental GC (#4979)
Grow the incremental GC partitions incrementally for EOP to reduce memory footprint. This wastes allocated memory for small canisters that do not explicitly allocate a lot of memory.

NB: before this change, `memory.grow()` was called in increments of 32/64 MiB (32 for classical persistence and 64 for Enhanced OP). Now, the growth happens at the granularity of `WASM_PAGE` size (64 KiB). A future performance improvement (if needed) can make allocations run at the granularity of several Wasm pages.