chore: automatically push accepted `test/bench` numbers on `x86_64-linux` (#5014)
Looks like our local build environments are too diverse to single out one platform for local testing. So I went with CI-side updating of the numbers, creating commits.

This will require folks to do either a `git pull` at times or a `git pull --rebase` if you added commits in the meantime. Trying to push without that will be rejected. I hope that is acceptable!