fix: issue #4309 (#5013)
Fixes #4309.

## The new Ansatz

This originally (in #5011 described approach) relied on penultimate `Con`s but I came up with a better idea.

The hiding regime is the same, but now I introduce a reverse mapping from the `normalize`d type to the monomorphised `id` (I do this per `Cons.name` plane). 

### What types look like internally

We want path compression to the _first seen_ `Con`, instead of expanding out completely:

``` mermaid
graph LR
    subgraph "Foo plane"
      A(Foo):::hide & B(Inner.Foo):::hide & C(Types.Foo):::hide & E(Types2.Foo):::hide --> D
      J(Other.Foo):::hide --> G
    end
    D("<code>Foo</code> in module
       (first seen)") -- normalise --o F["{ #bla }"]
    F == rev ===> D
    G("Some.Foo
         a.k.a. <code>Foo_1</code>") -- normalise --o H[Int]
    H == rev ===> G
    J -- normalise --o H
    classDef hide stroke-dasharray: 5 5
    linkStyle 6 stroke:green
    linkStyle 8 stroke:green
```
_Note_: not all `normalise` (i.e. —●) arrows drawn, to keep the diagram legible.

Preferably `Foo in module` should be the first `Foo` monomorphised, and even if there are others also monomorphised with different `id`s, those names should never be referred to by `VarT` when returned by `typ`. They are dashed-outline above and only appear in the `env` database of arrows (as arrowheads, and pointing to their first seen `Con`'s (in the plane), while that in turn points to the `normalize`d type — by bulleted arrow).

Whenever we normalise to a type, we follow the fat green edge (by `rev`) to arrive at a monomorphised `Con`'s `id`.

## Bird's eye algorithm

For any `Con`, look up the `env` arrow, if exists we are done. Else follow the `normalise` arrow and then the green reverse mapping to get the `VarT` of the canonical monomorphised `id`.

## Empirical results

I applied the new `moc` to the _Veronica_ project (a.k.a. VRC), which was previously scattered with `_1`-types in the Candid. Now there is only one, and that points to (subtly) different normalised types (~probably some leftover~ parametrised type — `type Profile<P>`). I am pretty happy with this result, and hopefully this will render Candid much more palatable to both humans and tools!