[VSC-26] Fix stamping in cache
Problem: If we have two files `A` and `B` such that `A` imports `B`, `B`
has a type `T`, and `A` has something like `type T = B.T`, then we'll
get an error message about non-productive types when time the
typechecker runs with a cache. This happens because `B.T` will be
stamped with `0` and cached, and `A.T` will be stamped with 1. But then
the next time we run the typechecker, we'll skip checking `B`, and the
stamp counter will be `0`. This will cause `A.T` to be stamped with `0`
as well. When the typechecker compares `A.T` and `B.T`, it will consider
those are the same and display the error message.

Solution: Consider the file name (besides the counter) when stamping and
call `Mo_types.Cons.session` for each file. Now stamps will be unique
again, because `eq` and `compare` will see that the files name are
different.