feat: defines a tree-sitter based Motoko mode for Emacs (#5043)
## Motivation

The existing `motoko-model.el` in this repo is based on `swift-mode`, which is a bit sad.

With Emacs 29+ we finally got built-in tree-sitter support, so now it's relatively easy to get nice syntax highlighting and auto-indentation behaviour. Most of the _actual work_ was done by this grammar author: https://github.com/polychromatist/tree-sitter-motoko

This is pretty bare-bones, but I intend on adding a few rules here or there as I spot things that don't highlight or indent nicely.

## Setup
Add this line to your config:
```lisp
(load "/path/to/motoko-ts-mode.el")
```

and run `treesit-install-language-grammar`, pass `motoko` as the grammar name and https://github.com/christoph-dfinity/tree-sitter-motoko as the grammar repo. The remaining prompts can stay as defaults.

Afterwards opening `.mo` files should give you highlighted files with basic indentation support.

![There are dozens of us](https://media4.giphy.com/media/v1.Y2lkPTc5MGI3NjExc2hraGg4cnlnMHdjMW5iNmF2YjR0cmU4N25idzFrY2I4ZGhwenF0NyZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/ReBGGJtbXrjbQJwByP/giphy.gif)