feat: type constructors: use per-file type constructor ids, elide type stamps when printing, hash stamps when needed (#5055)
* [VSC-26] Fix stamping in cache

Problem: If we have two files `A` and `B` such that `A` imports `B`, `B`
has a type `T`, and `A` has something like `type T = B.T`, then we'll
get an error message about non-productive types when time the
typechecker runs with a cache. This happens because `B.T` will be
stamped with `0` and cached, and `A.T` will be stamped with 1. But then
the next time we run the typechecker, we'll skip checking `B`, and the
stamp counter will be `0`. This will cause `A.T` to be stamped with `0`
as well. When the typechecker compares `A.T` and `B.T`, it will consider
those are the same and display the error message.

Solution: Consider the file name (besides the counter) when stamping and
call `Mo_types.Cons.session` for each file. Now stamps will be unique
again, because `eq` and `compare` will see that the files name are
different.

* display hash of stamp; don't show stamps in errors

* update some tests

* update drun tests

* update run tests

* Delete test/fail/log

* Delete test/cmp/files/version0.drun.most

* Delete test/cmp/files/version1.drun.most

* Delete test/cmp/files/version2.drun.most

* Delete test/cmp/files/version3.drun.most

* Delete test/fail/ok/abstract-msgs.tc.ok

* Delete test/cmp/files/version5.drun.most

* Delete test/cmp/files/version4.drun.most

* Revert "Delete test/fail/ok/abstract-msgs.tc.ok"

This reverts commit 76bcd6dc89d48ab9609870c1f7b3a9f46889b486.

* Update test/test-moc.js

* Update test/test-moc.js

* precompute hash; use to optimize compare

* update test output

* one that got away

---------

Co-authored-by: Heitor Toledo Lassarote de Paula <heitor.toledo@serokell.io>