[VSC-10] Add AST comparison tests
Problem: We should check that, when comparing the AST with and without
sources, they are the same (modulo the sources), including their type
representations.

Solution: Move `test_field_srcs` test from `mo_frontend` to `pipeline`
to be able to use `Pipeline` functions. Write a new test that loads
`.mo` files in some directories in `test` and checks that when
typechecking these files with `Mo_config.Flags.typechecker_combine_srcs`
off and on, they are still the same (ignoring source regions, as they
should be the only thing that is different.)