feat: introduce `await?` to short-circuit fulfilled `async` futures (#5215)
Fixes #5212.

Previously one had to use `await` (without the `*`) to force an `async` future to a value. The extended semantics implemented for `await?` will opt out of the the mandatory state commit coming with `await` and return the value from the future instantly if already fulfilled.

The intuition for the syntax is that this causes 0 or 1 commit points (the former when the future is already fulfilled in an earlier `await` or `await?`).

FYI: @mraszyk

TODO:
- [x] IR-interpreter
- [x] docs