fix: fix DWARF file indices in the `debug_line` section (#5281)
Instructions how to run it and how I've verified it when testing the motoko_debugger:
- Switch to this branch in the motoko repo and execute `nix develop` (this will put the locally build `moc` in the PATH)
- Have the `motoko-core` repo checkout on the branch `kamil/experiment-debugger-manual-test` next to this repo, go there: `cd ../motoko-core` (here we have the test files we would like to debug)
- Clone and [build](https://github.com/scalebit/motoko_debugger?tab=readme-ov-file#build-instructions) the `motoko_debugger`
- `make -C ../motoko/src moc && moc $(mops sources) -o MoDe.wasm -g --legacy-persistence test/List.test.mo && ~/motoko_debugger/target/release/debugger MoDe.wasm`. It builds the `moc` compiler and then compiles the problematic List test file, and finally runs the debugger on it. It should start and run to the end without problems now.

The debug line section in DWARF can be verified with
- `make -C ../motoko/src moc && moc $(mops sources) -o MoDe.wasm -g --legacy-persistence test/List.test.mo && llvm-dwarfdump --debug-line MoDe.wasm > out`
- open the out file in the editor and look for the `File` column, the file indices should not exceed the number of files

The only relevant changes are in the [src/wasm-exts/customModuleEncode.ml](https://github.com/dfinity/motoko/pull/5281/files#diff-95f6b3b28a82d084ba1f943362b8c74c7e767c290599ca8e44d4ca276acc8763) file. The other changes (compared with master) are due to an unrelated issue. I'll change the target of the branch to the branch related to that fix to limit the number of changes files in this PR. **Do not merge**

Fixes #4939