feat: enable unqualified import of types (#5056)
## Overview

Allows pattern matching on type fields of modules and objects.

```motoko
import { type Result } "mo:base/Result";

func parse(input : Text) : Result<Nat, Text> {
  ...
}

// More contrived example

func example() {
  class Example() {
    public type T = Text;
  };
  let ex : Example = Example();
  let { type T } = ex;
  let x : T = "Hello";
};
```

## Design considerations

The current implementation does not require specifying type parameters on the type field pattern:

```motoko
import { type Result } "...";
// As opposed to:
import { type Result<Ok, Err> } "...";
```

The more conservative choice would be to require to specify the parameters and their bounds, but so far I haven't come up with an example where this is a problem. We might want to require it anyway, as it would be a breaking change to require them at a future time.

At the moment the implementation only resolves these type fields patterns on `let`'s (and imports), and requires a restricted "value path" on the rhs of the let. We should be able to relax these restrictions in the future, but for now this lets us cover the important use-cases while keeping implementation complexity down. Relaxing these restrictions would be a non-breaking change.

## Testing

I added a few test cases to the test suite and made sure this continues to compile the "big package set".

Ran a couple manual tests for candid generation, and everything seemed to be fine.

## Reviewer notes

There are a few places with NOTE or TODO comments left, I'd like to discuss those during the review.