bugfix: issue 5336 (#5367)
Importing a `persistent` actor class would fail and report that the actor class must be declared `persistent` even though it was.

The problem is that the translation from ast to compilation unit and back again wasn't properly preserving the persistent annotation.

Also fixes another lurking bug in the parser that would be revealed when we turn on --default-persistent-actors.

Fixes #5336