feat: Weak References in Motoko (#5363)
First implementation of Weak References in Motoko with GC and type system support. The main functionality so far is to dereference a weak reference and check if the object pointed at by it is still live (i.e., wasn't garbage collected).

So far, the API exposes the following:
1. Create a weak reference to an object, via `weakRef = Prim.allocWeakRef(object)`.
2. Dereference a weak reference, via `object = Prim.getWeak()`.
3. Check if a target of a weak reference is still alive, via `Prim.isLive(weakRef)`.

This all works by adding a new type in the compiler and a new tag type in the RTS for tracking everything. The most non-intrusive change that allowed for achieving the aforementioned functionality was to add: (1) a new data structure that collects all the Weak References during marking; and (2) a new GC step immediately after marking that goes through this Weak Reference data structure and checks whether objects they point to are still live. If not, the weak reference pointer is updated to be NULL.

The functionality of the PR covers the difficult case of upgrades that may fall within a GC cycle by ensuring that the metadata layout is compatible even with adding a new data structure.