bugfix: issue-5283 (repro and fix) (#5412)
Fix for issue #5283.

The await translation would crash in obscure cases. Turns out its due to a bug in ir/rename.ml that wasn't properly distinguishing labels from identifiers when renaming them.

The counter-example is this code from the repro

```
                       label satset for (satset in [1,2,4].vals()) {
                          if (4==5) {
                            
                            break satset;
                          }
                        };
```

This uses a label and identifier with the same name, and unless kept distinct, after broken renaming,  would result in the inner break referencing the new identifier for (inner) satset rather than the new label for (outer) satset.

The fix is for renaming to use a map with a domain that distinguishes identifers from labels. 
Even better would be if we didn't use the same type (string, yuck) for identifiers and labels, but that is a bigger refactoring for later.