feat: implements contextual dot resolution
At a high level this allows writing a function call like
`List.concat(list1, list2)` as `list1.concat(list2)`, by finding the
`List.concat` function in the context using the inferred type of
`list1`. This allows for a more fluent style of using code defined
with data types and functions as opposed to classes and methods.