feat: implements contextual dot resolution (#5458)
At a high level this allows writing a function call like `List.concat(list1, list2)` as `list1.concat(list2)`, by finding the `List.concat` function in the context using the inferred type of `list1`. This allows for a more fluent style of using code defined with data types and functions as opposed to classes and methods.

This implements the design in https://github.com/dfinity/motoko/issues/2537 with its first two variations. It extends it slightly by not only taking the `Self` type of the module into account, but also checking compatibility with the first argument of the to-be-picked resolution.

### Todo
- [x] Decide if the `Self` module type is the way we want to go. I went with it over searching through every function in scope with the hope of keeping compile times in check.
- [x] Write up proper spec, by adjusting the design in https://github.com/dfinity/motoko/issues/2537 to the current implementation
- [x] Make a branch of `core` that uses this throughout to see if we like the impact (https://github.com/dfinity/motoko-core/pull/395)
- Documentation
  - [x] Language manual


### Future
- Integration with the language server. Don't want to block this PR on this item and I'm not sure what it involves, so I'll defer to the future.
- More informal prose with examples (For now just the Language manual has been updated)