chore: reduce the Mergify dependency (#5538)
It should be possible to instead use GitHub actions and the `gh` tool to
emulate the subset of functionality that we use Mergify for.

As communicated to the team:
> I have activated the “Enable auto-merge (squash)” button on PRs (with
the PR description as the default commit message) and tested it myself
successfully for #5533. Please start using it and report back if you
find something odd. We plan to phase-out Mergify this month and I’ll
prepare a PR to remove all mentions to it, as GH seems to provide the
same functionality that we need.

TODO:
- [ ] remove `.mergify.yml`
- [ ] compensate for `automerge-squash` in
`.github/workflows/update-flake-lock.yml`
- [x] work around `autoclose` in scheduled weekly runs
  - [ ] check `base=master` (previously in `.mergify.yml`)
  - [x] predicate this not on the step, but the top-level
- [ ] can we emulate approvals previously given by Mergify?
- [x] Clean `CI.md`
- [ ] Communicate the elimination to the IT team

Out of an abundance of caution I'll do the remaining work on a follow-up
PR.