fix: catch up with Candid spec changes (`reserved` ↛ `null`) (#5619)
Reproduce issue #5543. Fixes #5543.

The spec says that every Candid value is a subtype of `opt T`, and gives
`opt t` when `t <: T`. But the `null` type has no subtypes. I.e. we only
get out a `null : Null` when the binary Candid has a `0x7F` there. In
all other cases the coercion must fail, either trapping or indicating a
coercion failure to the caller (when recoverable). However `null : ?T`
can always be materialised when an argument is missing, similarly `null
: Null`. So we have to treat missing arguments and coercion failures
differently when the requested type is `Null` (or `null` respectively,
in Candid).

This PR corrects the reading of (e.g. `\x70`, `<reserved>`) in the
non-recoverable case and makes it a coercion failure.

Also adds improvements to the `candid-tests` runner to
- not ignore the case when a field type can be inferred, but not in the
record type (eliminating `M0164`, ignored test)
- adds `null` values to top-level argument tuples where the type
requests them
- adds `null` fields to record values (in argument sequences) where the
type requests them

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: pr-automation-bot-public[bot] <189003650+pr-automation-bot-public[bot]@users.noreply.github.com>