fix: repros and fixes for issues #5504, #5543 (#5505)
## Implement Martin's revised deserialization for Candid options

> This PR fixes coercion of values into nested optional types.

>Before the fix from this PR, `5 : nat64` would coerce to `opt 5 : opt
na64` (as expected) and `null : opt opt nat64` (unexpected). >This PR
fixes the unexpected result to be `opt opt 5 : opt opt nat64` (as
expected).

>Before the fix from this PR, `5 : int` would coerce to `null : opt opt
nat`. This PR fixes the result to be `opt null : opt opt nat`. The
>motivation for this change is to enable using optional types as
follows:
>- outer option characterizes if a value (of any type) is provided;
>- inner option expresses to optimistically deserialize the actual value
into the given type (without failing to deserialize and without losing
the information if a value of any type is provided).

>This PR also adds multiple test cases covering the change to coercion
(including coercion to nested optional types over `null` and
`reserved`).

c.f.
*  Rust fix https://github.com/dfinity/candid/pull/679
* Candid spec [PR](https://github.com/dfinity/candid/pull/678), in
particular this commit

https://github.com/dfinity/candid/pull/678/commits/90af30565382ac814ef954bb935cd369348e3604

## Furthermore (from #5619) fixes #5543.

The spec says that every Candid value is a subtype of `opt T`, and gives
`opt t` when `t <: T`. But the `null` type has no subtypes. I.e. we only
get out a `null : Null` when the binary Candid has a `0x7F` there. In
all other cases the coercion must fail, either trapping or indicating a
coercion failure to the caller (when recoverable). However `null : ?T`
can always be materialised when an argument is missing, similarly `null
: Null`. So we have to treat missing arguments and coercion failures
differently when the requested type is `Null` (or `null` respectively,
in Candid).

This PR corrects the reading of (e.g. `\x70`, `<reserved>`) in the
non-recoverable case and makes it a coercion failure.

Also adds improvements to the `candid-tests` runner to
- not ignore the case when a field type can be inferred, but not in the
record type (eliminating `M0164`, ignored test)
- adds `null` values to top-level argument tuples where the type
requests them
- adds `null` fields to record values (in argument sequences) where the
type requests them

------------------
TODO Tasks:
- [x] bump candid dependency to pick up revised test suite
- [x] It would be cool if the generated Motoko would be output on
`stderr` when `--diag` is passed to `candid-tests`
- [ ] potential simplifications suggested in #5619

---------

Co-authored-by: Gabor Greif <gabor@dfinity.org>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: pr-automation-bot-public[bot] <189003650+pr-automation-bot-public[bot]@users.noreply.github.com>