[VSC-34] Add more regions for references
Problem: When running `check_exp_field`, we should combine the regions
between the expression field and type field. Otherwise, some regions are
missing for the Visual Studio Code extension.

Solution: In `check_exp_field`, combine the sources of of the expression
and type fields. We do so using a union, replacing both the expression
and type sources.