feat: just make implicit a (contextual) keyword  (#5725)
If everyone agrees, I'd like to include this in Motoko 1.0.

This PR just makes `implicit` a keyword but lets you parse (implicit :
T) as before.

If we do this now, with 1.0, then adding syntax that needs it will
(likely) not be a breaking change in future (since we already have the
keyword).

This is essentially a cut-down version of #5659 (already approved by
Kamil), without the extra syntax and wildcard implicits removed.

---------

Co-authored-by: Christoph Hegemann <christoph.hegemann@dfinity.org>